Fixed issue with useEffect and re-rendering intro component. Added comments work that needs to be done
