+++ frontend/src/pages/Profile/ProfilePage.jsx
@@ -17,10 +17,10 @@ export const ProfilePage = () => {
     if (token) {
       getUserById(profile_id, token)
         .then((data) => {
-          // console.log(data);
           setProfileInfo(data.user);
           setToken(data.token);
           setUserID(data.user_id);
+
           if (profile_id === userID) {
             setProfileOwner(true);
           }
@@ -38,7 +38,12 @@ export const ProfilePage = () => {
     <div className="bg-slate-100 min-h-screen py-4">
       <div className="grid grid-cols-1 lg:grid-cols-2 gap-4 max-w-screen-lg mx-auto px-4">
         <div className="min-w-0 flex-auto">
-          <Intro profileInfo={profileInfo} profileOwner={profileOwner} token={token}/>
+          <Intro
+            profileInfo={profileInfo}
+            profileOwner={profileOwner}
+            token={token}
+            setToken={setToken}
+          />
           <div className="flex flex-col mx-auto my-3 pt-2 p-4 border shadow-sm rounded-lg bg-white">
             <h1 className="my-2 text-xl text-left font-bold tracking-tight text-gray-900">
               Friends
@@ -51,7 +56,7 @@ export const ProfilePage = () => {
           </div>
         </div>
         <div className="min-w-0 flex-auto">
-        <div className="flex flex-col mx-auto pt-2 p-4 border shadow-sm rounded-lg bg-white">
+          <div className="flex flex-col mx-auto pt-2 p-4 border shadow-sm rounded-lg bg-white">
             <h1 className="my-2 text-xl text-left font-bold tracking-tight text-gray-900">
               Posts
             </h1>
